REF: de-duplicate methods across nullable EA subclasses (#45380)
